Change how commits are handled.
I claim that whether a parser commits is a function of the
behavious of any sub-parsers that it used, and that passing
commit information around otherwise is not needed.

Simple (non-combinator) parsers all do not commit, except
for Commit.

Combinators commitness is a function of the commitness of
supplied sub-parsers.

This is at odds with edwinb's intention for commit to only
commit to some enclosing alternative, rather than all the way
to the root; but the global behaviour is what I want for now.